\name{s1}
\alias{s1}
\docType{data}
\title{
Example sequence data set
}
\description{
Example data set containing one single sequence
}
\usage{data(s1)}
\format{
  The format is:
 chr "s1"
}
\details{
A sequence object can be created with the dedicated \code{TraMineR} \code{\link{seqdef}} function. See example below. 
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(s1)
## maybe str(s1) ; plot(s1) ...
}
\keyword{datasets}
