\name{subtree-methods}
\docType{methods}
\alias{subtree}
\alias{subtree,PSTf-method}
\title{ Extract a subtree from a segmented PST }
\description{
 Extract a subtree from a segmented PST
}
\section{Methods}{
\describe{

\item{\code{signature(object = "PSTf")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{models}
